+++ superset-frontend/packages/superset-ui-core/src/style/index.tsx
@@ -61,7 +61,7 @@ const defaultTheme = {
     },
     primary: {
       base: '#0085FF',
-      dark1: '#1A85A0',
+      dark1: '#0085FF',
       dark2: '#156378',
       light1: '#79CADE',
       light2: '#A5DAE9',
