+++ superset-frontend/src/components/WalletConnector/WalletConnector.tsx
@@ -2,26 +2,33 @@ import React from 'react';
 import { css, styled } from '@superset-ui/core';
 import { ethers } from 'ethers';
 
+// eslint-disable-next-line theme-colors/no-literal-colors
 const StyledConnectWallet = styled.button`
   width: 118px;
   height: 25px;
-  background: #0085FF;
+  background: #0085ff;
   border-radius: 3px;
   font-size: 12px;
   font-weight: 400;
-  color: #FFFFFF;
+  color: #ffffff;
   border: None;
-`
+`;
 
-// TODO: Route to Flask API to get nonce
-async function getNonce(address: string) {
-    const data:object = await fetch(`/get_nonce?address=${address}`).then(res => res.json());
-    return data.nonce;
+interface WalletConnectorProps {
+  username: string;
 }
 
-async function handleWalletConnect() {
+function WalletConnector({ username }: WalletConnectorProps) {
+  async function getNonce(address: string) {
+    const data: object = await fetch(`/get_nonce?address=${address}`).then(
+      res => res.json(),
+    );
+    return data.nonce;
+  }
+
+  async function handleWalletConnect() {
     if (!window.ethereum) {
-        alert('Please install a crypto wallet');
+      alert('Please install a crypto wallet');
     }
     await window.ethereum.send('eth_requestAccounts');
     const provider = new ethers.providers.Web3Provider(window.ethereum);
@@ -29,17 +36,30 @@ async function handleWalletConnect() {
     const address = await signer.getAddress();
     const nonce = await getNonce(address);
     const signature = await signer.signMessage(nonce);
-    console.log(address, nonce, signature);
-    await fetch(`/login?address=${address}&signature=${signature}`)
+    await fetch(`/login?address=${address}&signature=${signature}`);
     location.reload();
-}
+  }
 
-function WalletConnector() {
+  function truncateText(text: string) {
+    if (text.length <= 10) {
+      return text;
+    }
+    return `${text.substr(0, 10)}...`;
+  }
+
   return (
-    <StyledConnectWallet onClick={handleWalletConnect}>
-        Connect Wallet
-    </StyledConnectWallet>
-  )
+    <div>
+      {username ? (
+        <StyledConnectWallet disabled>
+          {truncateText(username)}
+        </StyledConnectWallet>
+      ) : (
+        <StyledConnectWallet onClick={handleWalletConnect}>
+          Connect Wallet
+        </StyledConnectWallet>
+      )}
+    </div>
+  );
 }
 
-export default WalletConnector
+export default WalletConnector;
